+++ R/is.empty.stack.R
@@ -10,7 +10,8 @@
 #' @return 
 #' A logical, \code{TRUE} if \code{x} is empty. 
 #' 
-#' @seealso The generic \code{\link[base2]{is.empty}} in package \pkg{base2}. 
+#' @seealso The generic function \code{\link[base2]{is.empty}} 
+#' in package \pkg{base2}. 
 #' 
 #' @export
 #' 
@@ -18,5 +19,5 @@ is.empty.stack <-
 function(x)
 {
   stopifnot(is.stack(x))
-  size(x) == 0L
+  height(x) == 0L
 }