+++ R/new.R
@@ -6,9 +6,13 @@
 #' create 'First In First Out',  'Last In First Out', and 
 #' 'Not In or Never Out' stacks, respectively. 
 #' 
-#' @param max_size
-#' numeric. The maximum (possibly infinite) number of objects 
+#' @param max_height
+#' numeric. The maximum (infinite by default) number of objects 
 #' the stack can contain. 
+#' 
+#' @param max_size
+#' numeric. The maximum (infinite by default) size of  
+#' the stack, in octets.  
 #' 
 #' @param x
 #' An object to be tested or coerced. 
@@ -23,16 +27,19 @@
 #' an empty FIFO, LIFO, or NINO stack. 
 #' 
 #' @seealso 
-#' \code{\link[flifo]{push}}, \code{\link[flifo]{pop}}. 
+#' \code{\link[flifo]{push}}, 
+#' \code{\link[flifo]{pop}}. 
 #' 
 #' @export
 #' 
 fifo <- 
-function(max_size = Inf)
+function(max_height = Inf, 
+         max_size = Inf)
 {
   x <- list()
   class(x) <- c("fifo", "stack")
-  attr(x, "size") <- 0L
+  attr(x, "sizes") <- 0
+  attr(x, "max_height") <- max_height
   attr(x, "max_size") <- max_size
   x
 }
@@ -41,11 +48,13 @@ function(max_size = Inf)
 #' @rdname fifo
 #' 
 lifo <- 
-function(max_size = Inf)
+function(max_height = Inf, 
+         max_size = Inf)
 {
   x <- list()
   class(x) <- c("lifo", "stack")
-  attr(x, "size") <- 0L
+  attr(x, "sizes") <- 0
+  attr(x, "max_height") <- max_height
   attr(x, "max_size") <- max_size
   x
 }
@@ -55,11 +64,13 @@ function(max_size = Inf)
 #' @rdname fifo
 #' 
 nino <- 
-function(max_size = Inf)
+function(max_height = Inf, 
+         max_size = Inf)
 {
   x <- list()
   class(x) <- c("nino", "stack")
-  attr(x, "size") <- 0L
+  attr(x, "sizes") <- 0
+  attr(x, "max_height") <- max_height
   attr(x, "max_size") <- max_size
   x
 }