+++ NEWS.md
@@ -1,3 +1,10 @@
+# flifo 0.1.2
+
+* Every stack now has a 'max_height' and a 'max_size' attributes. 
+The 'height' function gives the current number of elements, 
+and the 'size' function gives the size of the stack in bytes. 
+
+
 # flifo 0.1.1
 
 * Now the print method works for NINO stacks.